Fixes gradle multi-project configuration. Completes Android ignored files configuration.
