+++ .gitignore
@@ -108,8 +108,21 @@ crashlytics.properties
 crashlytics-build.properties
 fabric.properties
 
+# Local configuration file (sdk path, etc)
+local.properties
+
+# Proguard folder generated by Eclipse
+proguard/
+
+# Keystore files
+*.jks
+
+# Freeline
+freeline.py
+freeline/
+freeline_project_description.json
+
 ### Other
 # Google Services (e.g. APIs or Firebase)
 google-services.json
-
