--- Mobile/app/src/main/java/com/s3project/mobilebooking/MainActivity.java
@@ -1,13 +1,0 @@
-package com.s3project.mobilebooking;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-    }
-}
