+++ backoffice/build.gradle
@@ -1,0 +1,13 @@
+apply plugin: 'java'
+apply plugin: 'scala'
+
+sourceCompatibility = 1.8
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    testCompile 'junit:junit:4.12'
+    compile 'org.scala-lang:scala-library:2.12.2'
+}
