--- gradle.properties
@@ -1,4 +1,0 @@
-theGroup=com.unibo
-theName=S3Project
-theVersion=1.0
-theSourceCompatibility=1.7
