+++ mobile/src/main/res/values/colors.xml
