+++ mobile/src/main/res/values/styles.xml
