+++ services/authentication/src/main/java/Main.java
@@ -1,0 +1,8 @@
+
+
+public class Main {
+
+    public static void main(String[] args) {
+        System.out.println("Hello World!");
+    }
+}
